refactor: user / auth 도메인 분리
