+++ src/main/java/org/example/namelesschamber/domain/auth/dto/request/LoginRequestDto.java
@@ -1,8 +1,8 @@
-package org.example.namelesschamber.domain.user.dto.request;
+package org.example.namelesschamber.domain.auth.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 
-public record SignupRequestDto(
+public record LoginRequestDto(
         @NotBlank
         String email,
         @NotBlank
