+++ src/main/java/org/example/namelesschamber/domain/auth/dto/request/ReissueRequestDto.java
@@ -1,4 +1,4 @@
-package org.example.namelesschamber.domain.user.dto.request;
+package org.example.namelesschamber.domain.auth.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
